+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Metrics endpoint for Prometheus integration
 * Search in list of targets using API & CLI
+* Allow filtering targets by arch value
 
 ## [0.5.0](https://git.rustybever.be/vieter-v/vieter/src/tag/0.5.0)